Fix Windows build
Summary:
Fix 2 issues that was breaking Windows build
1) double to size_t potential downcast warning
2) port_posix is not ready for windows, avoiding building hash_table_bench to
avoid build break

Test Plan: compile in Windoes and make check

Reviewers: sdong, andrewkr

Subscribers: andrewkr, dhruba, leveldb

Differential Revision: https://reviews.facebook.net/D59265

+++ utilities/persistent_cache/persistent_cache_tier.h
@@ -192,7 +192,8 @@ struct PersistentCacheOptions {
   // file size in order to avoid dead lock.
   size_t write_buffer_count() const {
     assert(write_buffer_size);
-    return (writer_qdepth + 1.2) * cache_file_size / write_buffer_size;
+    return static_cast<size_t>((writer_qdepth + 1.2) * cache_file_size /
+                               write_buffer_size);
   }
 
   // writer-dispatch-size
